legacy/backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-core
langchain-community
langchain-openai
langchain-ollama
langchain-qdrant
qdrant-client
pypdf
docx2txt
tiktoken
pyyaml
httpx
mcp
sse-starlette